+++ relayer/wallet/xrpl/types.go
@@ -4,10 +4,10 @@ type SignerPayload struct {
 	Account  string
 	OracleID uint64
 	Fee      string
-	Sequence uint64
+	Sequence uint32
 }
 
-func NewSignerPayload(account string, oracleID uint64, fee string, sequence uint64) SignerPayload {
+func NewSignerPayload(account string, oracleID uint64, fee string, sequence uint32) SignerPayload {
 	return SignerPayload{
 		Account:  account,
 		OracleID: oracleID,
